move demo icon to profile
